% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-lib.R
\name{checkFileDownload}
\alias{checkFileDownload}
\title{checkFileDownload}
\usage{
checkFileDownload(
  filepath,
  update.ts = Sys.time(),
  min.ts.diff = 12 * 60 * 60,
  min.size = Inf
)
}
\description{
checkFileDownload
}
