% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize-lib.R
\name{COVID19ARsummarizer}
\alias{COVID19ARsummarizer}
\title{COVID19ARsummarizer}
\description{
COVID19ARsummarizer

COVID19ARsummarizer
}
\author{
kenarab
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{COVID19ARsummarizer$new()}}
\item \href{#method-preprocess}{\code{COVID19ARsummarizer$preprocess()}}
\item \href{#method-generateJson}{\code{COVID19ARsummarizer$generateJson()}}
\item \href{#method-saveJson}{\code{COVID19ARsummarizer$saveJson()}}
\item \href{#method-clone}{\code{COVID19ARsummarizer$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19ARsummarizer$new(data.dir = getEnv("data_dir"))}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-preprocess"></a>}}
\if{latex}{\out{\hypertarget{method-preprocess}{}}}
\subsection{Method \code{preprocess()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19ARsummarizer$preprocess()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-generateJson"></a>}}
\if{latex}{\out{\hypertarget{method-generateJson}{}}}
\subsection{Method \code{generateJson()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19ARsummarizer$generateJson()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-saveJson"></a>}}
\if{latex}{\out{\hypertarget{method-saveJson}{}}}
\subsection{Method \code{saveJson()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19ARsummarizer$saveJson(
  dest.path = file.path(self$data.dir, paste("covid19ar_", as.character(self$date,
    format = "\%Y\%m\%d"), ".json", sep = ""))
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19ARsummarizer$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
