% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize-lib.R
\name{COVID19ARsummarizerCovidStats}
\alias{COVID19ARsummarizerCovidStats}
\title{COVID19ARsummarizerCovidStats}
\description{
COVID19ARsummarizerCovidStats

COVID19ARsummarizerCovidStats
}
\author{
kenarab
}
\section{Super class}{
\code{\link[COVID19AR:COVID19ARsummarizer]{COVID19AR::COVID19ARsummarizer}} -> \code{COVID19ARsummarizerCovidStats}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{COVID19ARsummarizerCovidStats$new()}}
\item \href{#method-preprocess}{\code{COVID19ARsummarizerCovidStats$preprocess()}}
\item \href{#method-aggregate}{\code{COVID19ARsummarizerCovidStats$aggregate()}}
\item \href{#method-generateOutputJson}{\code{COVID19ARsummarizerCovidStats$generateOutputJson()}}
\item \href{#method-clone}{\code{COVID19ARsummarizerCovidStats$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="COVID19AR" data-topic="COVID19ARsummarizer" data-id="generateJson">}\href{../../COVID19AR/html/COVID19ARsummarizer.html#method-generateJson}{\code{COVID19AR::COVID19ARsummarizer$generateJson()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="COVID19AR" data-topic="COVID19ARsummarizer" data-id="saveJson">}\href{../../COVID19AR/html/COVID19ARsummarizer.html#method-saveJson}{\code{COVID19AR::COVID19ARsummarizer$saveJson()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19ARsummarizerCovidStats$new(data.dir = getEnv("data_dir"))}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-preprocess"></a>}}
\if{latex}{\out{\hypertarget{method-preprocess}{}}}
\subsection{Method \code{preprocess()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19ARsummarizerCovidStats$preprocess(force.download = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-aggregate"></a>}}
\if{latex}{\out{\hypertarget{method-aggregate}{}}}
\subsection{Method \code{aggregate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19ARsummarizerCovidStats$aggregate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-generateOutputJson"></a>}}
\if{latex}{\out{\hypertarget{method-generateOutputJson}{}}}
\subsection{Method \code{generateOutputJson()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19ARsummarizerCovidStats$generateOutputJson(source = "covidstats")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19ARsummarizerCovidStats$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
