% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize-lib.R
\name{COVID19ARsummarizerMinsal}
\alias{COVID19ARsummarizerMinsal}
\title{COVID19ARsummarizerMinsal}
\description{
COVID19ARsummarizerMinsal

COVID19ARsummarizerMinsal
}
\author{
kenarab
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{COVID19ARsummarizerMinsal$new()}}
\item \href{#method-preprocessCasesMinsal}{\code{COVID19ARsummarizerMinsal$preprocessCasesMinsal()}}
\item \href{#method-preprocessVaccinesMinsal}{\code{COVID19ARsummarizerMinsal$preprocessVaccinesMinsal()}}
\item \href{#method-makeVaccinesTab}{\code{COVID19ARsummarizerMinsal$makeVaccinesTab()}}
\item \href{#method-preprocessMinSal}{\code{COVID19ARsummarizerMinsal$preprocessMinSal()}}
\item \href{#method-LoadDataFromCovidStats}{\code{COVID19ARsummarizerMinsal$LoadDataFromCovidStats()}}
\item \href{#method-generateOutputJson}{\code{COVID19ARsummarizerMinsal$generateOutputJson()}}
\item \href{#method-clone}{\code{COVID19ARsummarizerMinsal$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19ARsummarizerMinsal$new(
  data.dir = getEnv("data_dir"),
  cases.filename = "Covid19Casos.csv",
  vaccines.filename = "datos_nomivac_covid19.csv"
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-preprocessCasesMinsal"></a>}}
\if{latex}{\out{\hypertarget{method-preprocessCasesMinsal}{}}}
\subsection{Method \code{preprocessCasesMinsal()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19ARsummarizerMinsal$preprocessCasesMinsal(force.preprocess = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-preprocessVaccinesMinsal"></a>}}
\if{latex}{\out{\hypertarget{method-preprocessVaccinesMinsal}{}}}
\subsection{Method \code{preprocessVaccinesMinsal()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19ARsummarizerMinsal$preprocessVaccinesMinsal(force.preprocess = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-makeVaccinesTab"></a>}}
\if{latex}{\out{\hypertarget{method-makeVaccinesTab}{}}}
\subsection{Method \code{makeVaccinesTab()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19ARsummarizerMinsal$makeVaccinesTab(
  vaccines.agg.covidstats.df,
  dosis.prefix = "orden_dosis"
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-preprocessMinSal"></a>}}
\if{latex}{\out{\hypertarget{method-preprocessMinSal}{}}}
\subsection{Method \code{preprocessMinSal()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19ARsummarizerMinsal$preprocessMinSal(force.preprocess = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LoadDataFromCovidStats"></a>}}
\if{latex}{\out{\hypertarget{method-LoadDataFromCovidStats}{}}}
\subsection{Method \code{LoadDataFromCovidStats()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19ARsummarizerMinsal$LoadDataFromCovidStats(force.download = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-generateOutputJson"></a>}}
\if{latex}{\out{\hypertarget{method-generateOutputJson}{}}}
\subsection{Method \code{generateOutputJson()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19ARsummarizerMinsal$generateOutputJson(source = "covidstats")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19ARsummarizerMinsal$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
