% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eir-lib.R
\name{EcologicalInferenceCalvo}
\alias{EcologicalInferenceCalvo}
\title{EcologicalInferenceCalvo}
\description{
EcologicalInferenceCalvo

EcologicalInferenceCalvo
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{EcologicalInferenceCalvo$new()}}
\item \href{#method-wallTime}{\code{EcologicalInferenceCalvo$wallTime()}}
\item \href{#method-callDifp}{\code{EcologicalInferenceCalvo$callDifp()}}
\item \href{#method-paramsEstim}{\code{EcologicalInferenceCalvo$paramsEstim()}}
\item \href{#method-calcFractions}{\code{EcologicalInferenceCalvo$calcFractions()}}
\item \href{#method-paramsBoot}{\code{EcologicalInferenceCalvo$paramsBoot()}}
\item \href{#method-generateNormalizedOutput}{\code{EcologicalInferenceCalvo$generateNormalizedOutput()}}
\item \href{#method-makeSankeyDiagram}{\code{EcologicalInferenceCalvo$makeSankeyDiagram()}}
\item \href{#method-getFilename}{\code{EcologicalInferenceCalvo$getFilename()}}
\item \href{#method-generateOutputJSON}{\code{EcologicalInferenceCalvo$generateOutputJSON()}}
\item \href{#method-syncS3}{\code{EcologicalInferenceCalvo$syncS3()}}
\item \href{#method-getBetab}{\code{EcologicalInferenceCalvo$getBetab()}}
\item \href{#method-exportBetab}{\code{EcologicalInferenceCalvo$exportBetab()}}
\item \href{#method-runChile2021WithProcessor}{\code{EcologicalInferenceCalvo$runChile2021WithProcessor()}}
\item \href{#method-convertShares2Votes}{\code{EcologicalInferenceCalvo$convertShares2Votes()}}
\item \href{#method-convertVotes2Shares}{\code{EcologicalInferenceCalvo$convertVotes2Shares()}}
\item \href{#method-getSharesFields}{\code{EcologicalInferenceCalvo$getSharesFields()}}
\item \href{#method-runChile}{\code{EcologicalInferenceCalvo$runChile()}}
\item \href{#method-clone}{\code{EcologicalInferenceCalvo$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EcologicalInferenceCalvo$new(
  input.data,
  output.data,
  location.field = "COD_ZONA",
  seed = 143324
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-wallTime"></a>}}
\if{latex}{\out{\hypertarget{method-wallTime}{}}}
\subsection{Method \code{wallTime()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EcologicalInferenceCalvo$wallTime(expr)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-callDifp"></a>}}
\if{latex}{\out{\hypertarget{method-callDifp}{}}}
\subsection{Method \code{callDifp()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EcologicalInferenceCalvo$callDifp(p, mx, my, covar, nR, nC, nP, const)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-paramsEstim"></a>}}
\if{latex}{\out{\hypertarget{method-paramsEstim}{}}}
\subsection{Method \code{paramsEstim()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EcologicalInferenceCalvo$paramsEstim(
  data,
  x = -1,
  nR,
  nC,
  const = 0.001,
  parSeed = -1
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-calcFractions"></a>}}
\if{latex}{\out{\hypertarget{method-calcFractions}{}}}
\subsection{Method \code{calcFractions()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EcologicalInferenceCalvo$calcFractions(p, nR, nC, covar = F)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-paramsBoot"></a>}}
\if{latex}{\out{\hypertarget{method-paramsBoot}{}}}
\subsection{Method \code{paramsBoot()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EcologicalInferenceCalvo$paramsBoot(data, nR, nC, bootSamples)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-generateNormalizedOutput"></a>}}
\if{latex}{\out{\hypertarget{method-generateNormalizedOutput}{}}}
\subsection{Method \code{generateNormalizedOutput()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EcologicalInferenceCalvo$generateNormalizedOutput(indicator = "perc")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-makeSankeyDiagram"></a>}}
\if{latex}{\out{\hypertarget{method-makeSankeyDiagram}{}}}
\subsection{Method \code{makeSankeyDiagram()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EcologicalInferenceCalvo$makeSankeyDiagram(save.output = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getFilename"></a>}}
\if{latex}{\out{\hypertarget{method-getFilename}{}}}
\subsection{Method \code{getFilename()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EcologicalInferenceCalvo$getFilename(file.prefix, extension = "csv")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-generateOutputJSON"></a>}}
\if{latex}{\out{\hypertarget{method-generateOutputJSON}{}}}
\subsection{Method \code{generateOutputJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EcologicalInferenceCalvo$generateOutputJSON(
  save.output = TRUE,
  upload.s3 = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-syncS3"></a>}}
\if{latex}{\out{\hypertarget{method-syncS3}{}}}
\subsection{Method \code{syncS3()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EcologicalInferenceCalvo$syncS3()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getBetab"></a>}}
\if{latex}{\out{\hypertarget{method-getBetab}{}}}
\subsection{Method \code{getBetab()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EcologicalInferenceCalvo$getBetab()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-exportBetab"></a>}}
\if{latex}{\out{\hypertarget{method-exportBetab}{}}}
\subsection{Method \code{exportBetab()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EcologicalInferenceCalvo$exportBetab(
  output.folder = self$data.folder.s3,
  overwrite = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-runChile2021WithProcessor"></a>}}
\if{latex}{\out{\hypertarget{method-runChile2021WithProcessor}{}}}
\subsection{Method \code{runChile2021WithProcessor()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EcologicalInferenceCalvo$runChile2021WithProcessor(
  servel.processor,
  download.last.snapshot = TRUE,
  include.blancos = TRUE,
  include.ausentes = TRUE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-convertShares2Votes"></a>}}
\if{latex}{\out{\hypertarget{method-convertShares2Votes}{}}}
\subsection{Method \code{convertShares2Votes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EcologicalInferenceCalvo$convertShares2Votes(election.df)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-convertVotes2Shares"></a>}}
\if{latex}{\out{\hypertarget{method-convertVotes2Shares}{}}}
\subsection{Method \code{convertVotes2Shares()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EcologicalInferenceCalvo$convertVotes2Shares(election.df)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getSharesFields"></a>}}
\if{latex}{\out{\hypertarget{method-getSharesFields}{}}}
\subsection{Method \code{getSharesFields()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EcologicalInferenceCalvo$getSharesFields(election.fields)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-runChile"></a>}}
\if{latex}{\out{\hypertarget{method-runChile}{}}}
\subsection{Method \code{runChile()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EcologicalInferenceCalvo$runChile(
  include.blancos = TRUE,
  include.ausentes = TRUE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EcologicalInferenceCalvo$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
