% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COVID19AR-lib.R
\name{EdadCoder}
\alias{EdadCoder}
\title{EdadCoder}
\description{
EdadCoder

EdadCoder
}
\author{
kenarab
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{EdadCoder$new()}}
\item \href{#method-setupCoder}{\code{EdadCoder$setupCoder()}}
\item \href{#method-codeEdad}{\code{EdadCoder$codeEdad()}}
\item \href{#method-clone}{\code{EdadCoder$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EdadCoder$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setupCoder"></a>}}
\if{latex}{\out{\hypertarget{method-setupCoder}{}}}
\subsection{Method \code{setupCoder()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EdadCoder$setupCoder(
  age.breaks = c(0, 1, 10, 15, 20, 25, 30, 35, 40, 45, 50, 55, 60, 65, 70, 75, 80, 120),
  age.labels = c("0", "1-9", "10-14", "15-19", "20-24", "25-29", "30-34", "35-39",
    "40-44", "45-49", "50-54", "55-59", "60-64", "65-69", "70-74", "75-79", "80+")
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-codeEdad"></a>}}
\if{latex}{\out{\hypertarget{method-codeEdad}{}}}
\subsection{Method \code{codeEdad()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EdadCoder$codeEdad(edad)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EdadCoder$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
