% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize-lib.R
\name{COVID19ARsummarizerMinsal}
\alias{COVID19ARsummarizerMinsal}
\title{COVID19ARsummarizerMinsal}
\description{
COVID19ARsummarizerMinsal

COVID19ARsummarizerMinsal
}
\author{
kenarab
}
\section{Super class}{
\code{\link[COVID19AR:COVID19ARsummarizer]{COVID19AR::COVID19ARsummarizer}} -> \code{COVID19ARsummarizerMinsal}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{COVID19ARsummarizerMinsal$new()}}
\item \href{#method-preprocessCasesMinsal}{\code{COVID19ARsummarizerMinsal$preprocessCasesMinsal()}}
\item \href{#method-preprocessVaccinesMinsal}{\code{COVID19ARsummarizerMinsal$preprocessVaccinesMinsal()}}
\item \href{#method-setCurrentDate}{\code{COVID19ARsummarizerMinsal$setCurrentDate()}}
\item \href{#method-unzip}{\code{COVID19ARsummarizerMinsal$unzip()}}
\item \href{#method-preprocess}{\code{COVID19ARsummarizerMinsal$preprocess()}}
\item \href{#method-clone}{\code{COVID19ARsummarizerMinsal$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="COVID19AR" data-topic="COVID19ARsummarizer" data-id="aggregate">}\href{../../COVID19AR/html/COVID19ARsummarizer.html#method-aggregate}{\code{COVID19AR::COVID19ARsummarizer$aggregate()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="COVID19AR" data-topic="COVID19ARsummarizer" data-id="generateJson">}\href{../../COVID19AR/html/COVID19ARsummarizer.html#method-generateJson}{\code{COVID19AR::COVID19ARsummarizer$generateJson()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="COVID19AR" data-topic="COVID19ARsummarizer" data-id="saveJson">}\href{../../COVID19AR/html/COVID19ARsummarizer.html#method-saveJson}{\code{COVID19AR::COVID19ARsummarizer$saveJson()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19ARsummarizerMinsal$new(
  data.dir = getEnv("data_dir"),
  cases.filename = "Covid19Casos.csv",
  vaccines.filename = "datos_nomivac_covid19.csv",
  download.min.ts.diff = 19 * 60 * 60,
  ssl.version = "1.1"
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-preprocessCasesMinsal"></a>}}
\if{latex}{\out{\hypertarget{method-preprocessCasesMinsal}{}}}
\subsection{Method \code{preprocessCasesMinsal()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19ARsummarizerMinsal$preprocessCasesMinsal(force.preprocess = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-preprocessVaccinesMinsal"></a>}}
\if{latex}{\out{\hypertarget{method-preprocessVaccinesMinsal}{}}}
\subsection{Method \code{preprocessVaccinesMinsal()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19ARsummarizerMinsal$preprocessVaccinesMinsal(force.preprocess = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setCurrentDate"></a>}}
\if{latex}{\out{\hypertarget{method-setCurrentDate}{}}}
\subsection{Method \code{setCurrentDate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19ARsummarizerMinsal$setCurrentDate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-unzip"></a>}}
\if{latex}{\out{\hypertarget{method-unzip}{}}}
\subsection{Method \code{unzip()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19ARsummarizerMinsal$unzip(filepath)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-preprocess"></a>}}
\if{latex}{\out{\hypertarget{method-preprocess}{}}}
\subsection{Method \code{preprocess()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19ARsummarizerMinsal$preprocess(force.download = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19ARsummarizerMinsal$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
