% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eir-lib.R
\name{EcologicalInferenceGenerator}
\alias{EcologicalInferenceGenerator}
\title{EcologicalInferenceGenerator}
\description{
EcologicalInferenceGenerator

EcologicalInferenceGenerator
}
\author{
kenarab
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{EcologicalInferenceGenerator$new()}}
\item \href{#method-loadFilesFromMinSal}{\code{EcologicalInferenceGenerator$loadFilesFromMinSal()}}
\item \href{#method-preprocessMinSal}{\code{EcologicalInferenceGenerator$preprocessMinSal()}}
\item \href{#method-LoadDataFromCovidStats}{\code{EcologicalInferenceGenerator$LoadDataFromCovidStats()}}
\item \href{#method-merge}{\code{EcologicalInferenceGenerator$merge()}}
\item \href{#method-clone}{\code{EcologicalInferenceGenerator$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EcologicalInferenceGenerator$new(
  data.dir = "~/Downloads/",
  cases.filename = "Covid19Casos.csv",
  vaccines.filename = "datos_nomivac_covid19.csv"
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-loadFilesFromMinSal"></a>}}
\if{latex}{\out{\hypertarget{method-loadFilesFromMinSal}{}}}
\subsection{Method \code{loadFilesFromMinSal()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EcologicalInferenceGenerator$loadFilesFromMinSal(n.max = Inf)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-preprocessMinSal"></a>}}
\if{latex}{\out{\hypertarget{method-preprocessMinSal}{}}}
\subsection{Method \code{preprocessMinSal()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EcologicalInferenceGenerator$preprocessMinSal(n.max = 3e+06)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LoadDataFromCovidStats"></a>}}
\if{latex}{\out{\hypertarget{method-LoadDataFromCovidStats}{}}}
\subsection{Method \code{LoadDataFromCovidStats()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EcologicalInferenceGenerator$LoadDataFromCovidStats(force.download = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-merge"></a>}}
\if{latex}{\out{\hypertarget{method-merge}{}}}
\subsection{Method \code{merge()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EcologicalInferenceGenerator$merge()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EcologicalInferenceGenerator$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
